backend/pyproject.toml
[project]
name = "ops-backend"
version = "0.1.0"
description = "OPS backend for Organic Population Simulation and scenario forecasting."
requires-python = ">=3.11,<3.12"
license = { text = "AGPL-3.0" }
authors = [
    { name = "OPS Team" }
]

dependencies = [
    # Core framework
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "gunicorn>=23.0.0",
    
    # LLM-related
    "openai>=1.0.0",
    "supabase>=2.0.0",

    # Zep Cloud
    "zep-cloud==3.13.0",
    
    # OASIS social media simulation
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    
    # File handling
    "PyMuPDF>=1.24.0",
    # Encoding detection (supports text files with non-UTF-8 encodings)
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    
    # Utility libraries
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
